Refactor code, add tests
