+++ tests/Exception/InvalidMiddlewareReturnTypeExceptionTest.php
@@ -4,73 +4,16 @@ declare(strict_types=1);
 
 namespace Yiisoft\Middleware\Dispatcher\Tests\Exception;
 
-use PHPUnit\Framework\TestCase;
+use Exception\AbstractInvalidMiddlewareExceptionTest;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use stdClass;
+use Throwable;
 use Yiisoft\Middleware\Dispatcher\Exception\InvalidMiddlewareReturnTypeException;
 use Yiisoft\Middleware\Dispatcher\Helper\ResponseHelper;
-use Yiisoft\Middleware\Dispatcher\InvalidMiddlewareDefinitionException;
-use Yiisoft\Middleware\Dispatcher\Tests\Support\TestController;
 
-final class InvalidMiddlewareReturnTypeExceptionTest extends TestCase
+final class InvalidMiddlewareReturnTypeExceptionTest extends AbstractInvalidMiddlewareExceptionTest
 {
-    public function dataBase(): array
-    {
-        return [
-            [
-                'test',
-                '"test"',
-                'Class `test` not found',
-            ],
-            [
-                TestController::class,
-                '"Yiisoft\Middleware\Dispatcher\Tests\Support\TestController"',
-                'Class `Yiisoft\Middleware\Dispatcher\Tests\Support\TestController` exists, but does not implement',
-            ],
-            [
-                new TestController(),
-                'an instance of `Yiisoft\Middleware\Dispatcher\Tests\Support\TestController`',
-                'Related links',
-            ],
-            [
-                [TestController::class, 'notExistsAction'],
-                '["Yiisoft\Middleware\Dispatcher\Tests\Support\TestController", "notExistsAction"]',
-                'Try adding `notExistsAction()` action to ' .
-                '`Yiisoft\Middleware\Dispatcher\Tests\Support\TestController` controller:',
-            ],
-            [
-                ['class' => TestController::class, 'index'],
-                '["class" => "Yiisoft\Middleware\Dispatcher\Tests\Support\TestController", "index"]',
-                null,
-            ],
-            [
-                ['object' => TestController::class, 'index'],
-                '["object" => "Yiisoft\Middleware\Dispatcher\Tests\Support\TestController", "index"]',
-                'You may have an error in array definition. Array definition validation result',
-            ],
-            [
-                ['class' => TestController::class],
-                '["class" => "Yiisoft\Middleware\Dispatcher\Tests\Support\TestController"]',
-                'Array definition is valid, ' .
-                'class `Yiisoft\Middleware\Dispatcher\Tests\Support\TestController` exists, ' .
-                'but does not implement `Psr\Http\Server\MiddlewareInterface`.',
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider dataBase
-     */
-    public function testBase(mixed $definition, string $expectedMessage, ?string $expectedSolution): void
-    {
-        $exception = new InvalidMiddlewareDefinitionException($definition);
-        self::assertStringEndsWith('. Got ' . $expectedMessage . '.', $exception->getMessage());
-        if ($expectedSolution !== null) {
-            self::assertStringContainsString($expectedSolution, $exception->getSolution());
-        }
-    }
-
     public function dataInvalidReturnType(): array
     {
         return [
@@ -102,7 +45,7 @@ final class InvalidMiddlewareReturnTypeExceptionTest extends TestCase
         );
     }
 
-    public static function dataProviderName()
+    public static function dataProviderName(): iterable
     {
         yield 'null' => [null];
         yield 'string' => ['test'];
@@ -122,5 +65,10 @@ final class InvalidMiddlewareReturnTypeExceptionTest extends TestCase
             sprintf('Invalid middleware result type %s', get_debug_type($result)),
             $exception->getName()
         );
+    }
+
+    protected function createException(mixed $definition): Throwable
+    {
+        return new InvalidMiddlewareReturnTypeException($definition, new stdClass());
     }
 }
